refactor key generation FE code
