+++ kinode/src/register-ui/src/pages/RegisterKnsName.tsx
@@ -2,13 +2,13 @@ import { useState, useEffect, FormEvent, useCallback } from "react";
 import { hooks } from "../connectors/metamask";
 import { Link, useNavigate } from "react-router-dom";
 import { toDNSWireFormat } from "../utils/dnsWire";
-import { BytesLike, utils } from 'ethers';
+import { utils } from 'ethers';
 import EnterKnsName from "../components/EnterKnsName";
 import Loader from "../components/Loader";
 import KinodeHeader from "../components/KnsHeader";
-import { NetworkingInfo, PageProps } from "../lib/types";
-import { ipToNumber } from "../utils/ipToNumber";
-import { getNetworkName, setChain } from "../utils/chain";
+import { PageProps } from "../lib/types";
+
+import { generateNetworkingKeys, getNetworkName } from "../utils/chain";
 import DirectCheckbox from "../components/DirectCheckbox";
 import { Tooltip } from "../components/Tooltip";
 
@@ -58,62 +58,22 @@ function RegisterKnsName({
 
     if (!provider || !kns) return openConnect()
 
+    setLoading('Please confirm the transaction in your wallet');
     try {
-      setLoading('Please confirm the transaction in your wallet');
-      let networkingInfoResponse;
-      try {
-        const response = await fetch('/generate-networking-info', { method: 'POST' });
-        if (!response.ok) {
-          throw new Error(`HTTP error! status: ${response.status}`);
-        }
-        networkingInfoResponse = await response.json() as NetworkingInfo;
-      } catch (error) {
-        console.error('Failed to fetch networking info:', error);
-        throw error;
-      }
+      const nameToSet = utils.namehash(`${name}.os`);
 
-      const {
-        networking_key,
-        routing: {
-          Both: {
-            ip: ip_address,
-            ports: { ws: ws_port, tcp: tcp_port },
-            routers: allowed_routers
-          }
-        }
-      } = networkingInfoResponse;
-
-      const ipAddress = ipToNumber(ip_address)
-
-      setNetworkingKey(networking_key)
-      setIpAddress(ipAddress)
-      setWsPort(ws_port || 0)
-      setTcpPort(tcp_port || 0)
-      setRouters(allowed_routers)
-
-      const data: BytesLike[] = [
-        direct
-          ? (await kns.populateTransaction.setAllIp(
-            utils.namehash(`${name}.os`),
-            ipAddress,
-            ws_port || 0,  // ws
-            0,             // wt
-            tcp_port || 0, // tcp
-            0              // udp
-          )).data!
-          : (await kns.populateTransaction.setRouters
-            (utils.namehash(`${name}.os`), allowed_routers.map(x => utils.namehash(x)))).data!,
-        (await kns.populateTransaction.setKey(utils.namehash(`${name}.os`), networking_key)).data!
-      ]
-
-      setLoading('Please confirm the transaction in your wallet');
-
-      try {
-        await setChain(nodeChainId);
-      } catch (error) {
-        window.alert(`You must connect to the ${chainName} network to continue. Please connect and try again.`);
-        throw new Error(`${chainName} not set`)
-      }
+      const data = await generateNetworkingKeys({
+        direct,
+        kns,
+        nodeChainId,
+        chainName,
+        nameToSet,
+        setNetworkingKey,
+        setIpAddress,
+        setWsPort,
+        setTcpPort,
+        setRouters,
+      });
 
       const dnsFormat = toDNSWireFormat(`${name}.os`);
       const tx = await dotOs?.register(
