+++ kinode/src/register-ui/src/pages/ResetNode.tsx
@@ -6,13 +6,12 @@ import {
 } from "react";
 import { hooks } from "../connectors/metamask";
 import { Link, useNavigate } from "react-router-dom";
-import { namehash } from "ethers/lib/utils";
 import Loader from "../components/Loader";
 import KinodeHeader from "../components/KnsHeader";
-import { NetworkingInfo, PageProps } from "../lib/types";
-import { ipToNumber } from "../utils/ipToNumber";
-import { getNetworkName, setChain } from "../utils/chain";
+import { PageProps } from "../lib/types";
+import { generateNetworkingKeys, getNetworkName } from "../utils/chain";
 import DirectCheckbox from "../components/DirectCheckbox";
+import { namehash } from "@ethersproject/hash";
 
 const { useProvider } = hooks;
 
@@ -52,64 +51,20 @@ function ResetNode({
 
             if (!provider) return openConnect();
 
+            setLoading("Please confirm the transaction in your wallet");
             try {
-                setLoading("Please confirm the transaction in your wallet");
-
-                const {
-                    networking_key,
-                    routing: {
-                        Both: {
-                            ip: ip_address,
-                            ports: { ws: ws_port, tcp: tcp_port },
-                            routers: allowed_routers
-                        }
-                    }
-                } = (await fetch("/generate-networking-info", { method: "POST" }).then(
-                    (res) => res.json()
-                )) as NetworkingInfo;
-
-                const ipAddress = ipToNumber(ip_address);
-
-                setNetworkingKey(networking_key);
-                setIpAddress(ipAddress);
-                setWsPort(ws_port || 0);
-                setTcpPort(tcp_port || 0);
-                setRouters(allowed_routers);
-
-                const data = [
-                    direct
-                        ? (
-                            await kns.populateTransaction.setAllIp(
-                                namehash(knsName),
-                                ipAddress,
-                                ws_port || 0,  // ws
-                                0,             // wt
-                                tcp_port || 0, // tcp
-                                0              // udp
-                            )
-                        ).data!
-                        : (
-                            await kns.populateTransaction.setRouters(
-                                namehash(knsName),
-                                allowed_routers.map((x) => namehash(x))
-                            )
-                        ).data!,
-                    (
-                        await kns.populateTransaction.setKey(
-                            namehash(knsName),
-                            networking_key
-                        )
-                    ).data!,
-                ];
-
-                try {
-                    await setChain(nodeChainId);
-                } catch (error) {
-                    window.alert(
-                        `You must connect to the ${chainName} network to continue. Please connect and try again.`
-                    );
-                    throw new Error(`${chainName} not set`);
-                }
+                const data = await generateNetworkingKeys({
+                    direct,
+                    kns,
+                    nodeChainId,
+                    chainName,
+                    nameToSet: namehash(knsName),
+                    setNetworkingKey,
+                    setIpAddress,
+                    setWsPort,
+                    setTcpPort,
+                    setRouters,
+                });
 
                 const tx = await kns.multicall(data);
 
@@ -118,12 +73,12 @@ function ResetNode({
                 await tx.wait();
 
                 setReset(true);
-                setLoading("");
                 setDirect(direct);
                 navigate("/set-password");
             } catch {
-                setLoading("");
                 alert("An error occurred, please try again.");
+            } finally {
+                setLoading("");
             }
         },
         [
